Submission/requirements.txt
chromadb
langchain
tqdm
llama-cpp-python
transformers
numpy
langchain-community